refactor: reorganise animation lab
